pyproject.toml
[tool.poetry]
name = "diputado-distrito"
version = "0.1.0"
description = ""
authors = ["Nacho Platon <iplaton@geoblink.com>"]

[tool.poetry.dependencies]
python = "^3.6.8"
jupyterlab = "^2.2"
geopandas = "^0.8.1"
keplergl = "^0.2.1"
psycopg2-binary = "^2.8"
descartes = "^1.1"
gerrychain = "^0.2.12"
tqdm = "^4.52"
Rtree = "^0.9.4"
SQLAlchemy = "^1.3"
pyarrow = "^3.0"
unidecode = "^1.2"
openpyxl = "^3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"